Renamve QueryAnalysis to Resolve
